feat: refactor all code
