+++ setup.py
@@ -1,19 +1,18 @@
-# coding: utf-8
-from setuptools import setup
+"""Install the package."""
+from setuptools import setup  # type: ignore
 
 
 setup(
-    name='pylint-relative-imports',
-    version='0.0.4',
-    package_dir={'': 'src'},
-    packages=['pylint_relative_imports'],
-    url='https://github.com/kubikvid/pylint-relative-imports',
+    name='pylint_absolute_imports',
+    version='1.0',
+    packages=['pylint_absolute_imports'],
+    url='https://github.com/quentinn42/pylint_absolute_imports',
     license='MIT License',
-    author='Andrey Chernykh',
-    author_email='moonquiz@ya.ru',
+    author='Quentin Lieumont',
+    author_email='quentin@lieumont.fr',
     description=(
-        'Pylint plugin which adds linter error for '
-        'relative imports from the root of the project'
+        'Pylint plugin which adds linter error for relatives imports.'
+        'Read more about imports at https://peps.python.org/pep-0008/#imports'
     ),
     classifiers=(
         'Development Status :: 4 - Beta',
